# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação API controle de acesso Laravel",
        "description": "Documentação com todos os endpoints necessários para a correta utilização da aplicação.",
        "contact": {
            "email": "assuncaofilipe97@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/app/user/profile": {
            "post": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Cadastro de dados pessoais do usuário da API",
                "description": "Cadastra os dados pessoais do usuário",
                "operationId": "74e3d896e45f9212a2e6d2fb98a6a5e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "cpf": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "register-profile": {
                                    "summary": "Exemplo cpf sem separador",
                                    "value": {
                                        "first_name": "Olávo",
                                        "last_name": "Sales",
                                        "cpf": "61906713065"
                                    }
                                },
                                "register-profile1": {
                                    "summary": "Exemplo cpf com separador",
                                    "value": {
                                        "first_name": "Olávo",
                                        "last_name": "Sales",
                                        "cpf": "619.067.130-65"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cadastro feito com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/profileResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "email": {
                                                    "type": "string"
                                                },
                                                "password": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "examples": {
                                    "obrigatório": {
                                        "summary": "Campos obrigatórios",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "email": [
                                                    "O campo primeiro nome é obrigatório."
                                                ],
                                                "password": [
                                                    "O campo sobrenome é obrigatório."
                                                ],
                                                "terms_of_use": [
                                                    "O campo cpf é obrigatório."
                                                ]
                                            }
                                        }
                                    },
                                    "comfirmação": {
                                        "summary": "Campos inválidos",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "first_name": [
                                                    "O campo primeiro nome só pode conter letras.",
                                                    "O campo primeiro nome não pode ser superior a 50 caracteres."
                                                ],
                                                "last_name": [
                                                    "O campo sobrenome só pode conter letras.",
                                                    "O campo sobrenome não pode ser superior a 100 caracteres."
                                                ],
                                                "cpf": [
                                                    "O campo cpf já está sendo utilizado.",
                                                    "CPF inválido"
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/forbidden"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "profileResponse": {
                "properties": {
                    "success": {
                        "type": "string"
                    },
                    "data": {
                        "properties": {
                            "first_name": {
                                "type": "string"
                            },
                            "last_name": {
                                "type": "string"
                            },
                            "cpf": {
                                "type": "string"
                            },
                            "uuid": {
                                "type": "string"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date"
                            },
                            "id": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "success": "true",
                    "data": {
                        "first_name": "Olávo",
                        "last_name": "Sales",
                        "cpf": "61906713065",
                        "uuid": "250c9e0e-65df-4658-ad08-4cfe6d8abbf9",
                        "updated_at": "2022-02-10T17:09:56.000000Z",
                        "created_at": "2022-02-10T17:09:56.000000Z",
                        "id": 2
                    }
                }
            },
...
        },
        "responses": {
            "forbidden": {
                "description": "Token de autorização não encontrado / Token expirado",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "string"
                                },
                                "data": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        },
                        "examples": {
                            "Token de autorização não encontrado": {
                                "summary": "Token de autorização não encontrado",
                                "value": {
                                    "success": "false",
                                    "data": {
                                        "message": "Token de autorização não encontrado"
                                    }
                                }
                            },
                            "Token expirado": {
                                "summary": "Token expirado",
                                "value": {
                                    "success": "false",
                                    "data": {
                                        "message": "Token expirado"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "unauthorized": {
                "description": "Token inválido",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "string"
                                },
                                "data": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "success": "false",
                            "data": {
                                "message": "Token inválido"
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}